Fixed destroy product without image and change Observers to unique UrlObserver

+++ app/app/Http/Controllers/Admin/ProductController.php
@@ -68,7 +68,7 @@ class ProductController extends Controller
     public function destroy($id) 
     {
         $product = $this->repository->getById($id);
-        if(Storage::exists($product->image)) {
+        if(!empty($product->image) && Storage::exists($product->image)) {
             Storage::delete($product->image);
         }
         $this->repository->delete($id);
